requirements.txt
playwright==1.40.0
pandas==2.1.3
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
geopy==2.3.0
phonenumbers==8.13.0
pillow==10.1.0
lxml==4.9.3
aiohttp==3.9.1
gspread==5.11.0
google-auth-oauthlib==1.1.0
google-auth==2.25.2